Removing dependency on Docker to ran load database script.

+++ docker/load_database_backup.sh
@@ -27,6 +27,10 @@ while [[ $# -gt 0 ]]; do
             prod="true"
             shift
             ;;
+        --nodocker)      ## flag to run this outside docker environment
+            dock="false"
+            shift
+            ;;
         --postgres)    ## path to the postgres bin directory
             pgpath="$2"
             shift
@@ -83,48 +87,53 @@ fi
 # Take down the entire docker-compose project, including the network and volumes
 # then build a new postgresql configuration using the specified one as a base.
 #-------------------------------------------------------------------------------
-docker-compose down -v
-if [[ ! -e .env ]]; then
-    cp default.env .env
+
+echo -n 'value of dock '$dock
+if [[ -z $dock ]]; then
+  docker-compose down -v
+  if [[ ! -e .env ]]; then
+      cp default.env .env
+  fi
+  if [[ .env =~ "PG_CONF_FILE=(.*)" ]]; then
+      conf="${BASH_REMATCH[1]}"
+  else
+      conf="./postgres/postgresql.conf"
+  fi
+  sed -e "s/^.*fsync *=.*$/fsync = off/" \
+      -e "s/^.*synchronous_commit *=.*$/synchronous_commit = off/" \
+      -e "s/^.*wal_level *=.*$/wal_level = minimal/" \
+      -e "s/^.*full_page_writes *=.*$/full_page_writes = off/" \
+      -e "s/^.*checkpoint_segments *=.*$/checkpoint_segments = 64/" \
+      -e "s/^.*max_wal_senders *=.*$/max_wal_senders = 0/" \
+      -e "s/^.*wal_keep_segments *=.*$/wal_keep_segments = 0/" \
+      -e "s/^.*archive_mode *=.*$/archive_mode = off/" \
+      -e "s/^.*autovacuum *=.*$/autovacuum = off/" \
+      -e "s/^.*log_min_duration_statement *=.*$/log_min_duration_statement = -1/" \
+      -e "s/^.*log_checkpoints *=.*$/log_checkpoints = off/" \
+      -e "s/^.*log_connections *=.*$/log_connections = off/" \
+      -e "s/^.*log_disconnections *=.*$/log_disconnections = off/" \
+      -e "s/^.*log_duration *=.*$/log_duration = off/" \
+      -e "s/^.*log_hostname *=.*$/log_hostname = off/" \
+      -e "s/^.*log_lock_waits *=.*$/log_lock_waits = off/" \
+      -e "s/^.*log_statement *=.*$/log_statement = 'none'/" \
+      -e "s/^.*log_temp_files *=.*$/log_temp_files = -1/" \
+      -e "s/^.*effective_cache_size *=.*$/effective_cache_size = 2GB/" \
+      -e "s/^.*shared_buffers *=.*$/shared_buffers = 128MB/" \
+      -e "s/^.*maintenance_work_mem *=.*$/maintenance_work_mem = 1GB/" \
+      $conf > $tmpdir/pg_restore.conf
+  export PG_CONF_FILE=$tmpdir/pg_restore.conf
+  docker-compose up -d postgres
+  pgport=$(docker-compose port postgres 5432)
 fi
-if [[ .env =~ "PG_CONF_FILE=(.*)" ]]; then
-    conf="${BASH_REMATCH[1]}"
-else
-    conf="./postgres/postgresql.conf"
-fi
-sed -e "s/^.*fsync *=.*$/fsync = off/" \
-    -e "s/^.*synchronous_commit *=.*$/synchronous_commit = off/" \
-    -e "s/^.*wal_level *=.*$/wal_level = minimal/" \
-    -e "s/^.*full_page_writes *=.*$/full_page_writes = off/" \
-    -e "s/^.*checkpoint_segments *=.*$/checkpoint_segments = 64/" \
-    -e "s/^.*max_wal_senders *=.*$/max_wal_senders = 0/" \
-    -e "s/^.*wal_keep_segments *=.*$/wal_keep_segments = 0/" \
-    -e "s/^.*archive_mode *=.*$/archive_mode = off/" \
-    -e "s/^.*autovacuum *=.*$/autovacuum = off/" \
-    -e "s/^.*log_min_duration_statement *=.*$/log_min_duration_statement = -1/" \
-    -e "s/^.*log_checkpoints *=.*$/log_checkpoints = off/" \
-    -e "s/^.*log_connections *=.*$/log_connections = off/" \
-    -e "s/^.*log_disconnections *=.*$/log_disconnections = off/" \
-    -e "s/^.*log_duration *=.*$/log_duration = off/" \
-    -e "s/^.*log_hostname *=.*$/log_hostname = off/" \
-    -e "s/^.*log_lock_waits *=.*$/log_lock_waits = off/" \
-    -e "s/^.*log_statement *=.*$/log_statement = 'none'/" \
-    -e "s/^.*log_temp_files *=.*$/log_temp_files = -1/" \
-    -e "s/^.*effective_cache_size *=.*$/effective_cache_size = 2GB/" \
-    -e "s/^.*shared_buffers *=.*$/shared_buffers = 128MB/" \
-    -e "s/^.*maintenance_work_mem *=.*$/maintenance_work_mem = 1GB/" \
-    $conf > $tmpdir/pg_restore.conf
-export PG_CONF_FILE=$tmpdir/pg_restore.conf
-docker-compose up -d postgres
-pgport=$(docker-compose port postgres 5432)
 
 #-------------------------------------------------------------------------------
 # Wait for the postgres instance to start accepting connections
 #-------------------------------------------------------------------------------
-echo -n 'Waiting for postgres to start ... '
-docker-compose exec postgres /bin/bash -c 'count=0;while [ $count -lt 120 ]; do if psql -U postgres -c "\l" &>/dev/null; then sleep 3; break; fi; sleep 1; let count=count+1; done;' &>/dev/null
-echo -e '\033[0;32mdone\033[0m'
-
+if [[ -z $dock ]]; then
+  echo -n 'Waiting for postgres to start ... '
+  docker-compose exec postgres /bin/bash -c 'count=0;while [ $count -lt 120 ]; do if psql -U postgres -c "\l" &>/dev/null; then sleep 3; break; fi; sleep 1; let count=count+1; done;' &>/dev/null
+  echo -e '\033[0;32mdone\033[0m'
+fi
 #-------------------------------------------------------------------------------
 # If the user did not provide a path to an existing dump file, secure copy the
 # latest daily from the EHR production server's backup folder
@@ -138,11 +147,19 @@ fi
 #-------------------------------------------------------------------------------
 # Drop and recreate the labkey database and the various roles that we use
 #-------------------------------------------------------------------------------
-echo -n 'Preparing database and roles ... '
-docker-compose exec postgres psql -U postgres -c "drop database if exists ${dbname};" &>/dev/null
-docker-compose exec postgres psql -U postgres -c "create database ${dbname};" &>/dev/null
-docker-compose exec postgres psql -U postgres -c 'drop role if exists labkey; create role labkey superuser; drop role if exists doconnor; create role doconnor superuser; drop role if exists oconnor; create role oconnor superuser; drop role if exists oconnorlab; create role oconnorlab superuser; drop role if exists sconnor; create role sconnor superuser; drop role if exists soconnorlab; create role soconnorlab superuser; drop role if exists soconnor_lab; create role soconnor_lab superuser;' &>/dev/null
-echo -e '\033[0;32mdone\033[0m'
+if [[ -z $dock ]]; then
+  echo -n 'Preparing database and roles ... '
+  docker-compose exec postgres psql -U postgres -c "drop database if exists ${dbname};" &>/dev/null
+  docker-compose exec postgres psql -U postgres -c "create database ${dbname};" &>/dev/null
+  docker-compose exec postgres psql -U postgres -c 'drop role if exists labkey; create role labkey superuser; drop role if exists doconnor; create role doconnor superuser; drop role if exists oconnor; create role oconnor superuser; drop role if exists oconnorlab; create role oconnorlab superuser; drop role if exists sconnor; create role sconnor superuser; drop role if exists soconnorlab; create role soconnorlab superuser; drop role if exists soconnor_lab; create role soconnor_lab superuser;' &>/dev/null
+  echo -e '\033[0;32mdone\033[0m'
+else
+  echo -n 'Preparing database and roles ... '
+  ${pgpath}psql -U postgres -c "drop database if exists ${dbname};" &>/dev/null
+  ${pgpath}psql -U postgres -c "create database ${dbname};" &>/dev/null
+  ${pgpath}psql -U postgres -c 'drop role if exists labkey; create role labkey superuser; drop role if exists doconnor; create role doconnor superuser; drop role if exists oconnor; create role oconnor superuser; drop role if exists oconnorlab; create role oconnorlab superuser; drop role if exists sconnor; create role sconnor superuser; drop role if exists soconnorlab; create role soconnorlab superuser; drop role if exists soconnor_lab; create role soconnor_lab superuser;' &>/dev/null
+  echo -e '\033[0;32mdone\033[0m'
+fi
 
 #-------------------------------------------------------------------------------
 # Actually restore the database, using a background proc so we can track progress
@@ -196,6 +213,8 @@ fi
 # Tear down and re-start the docker-compose environment using the 'regular'
 # postgresql configuration rather than the 'restore' one
 #-------------------------------------------------------------------------------
-docker-compose down -v
-unset PG_CONF_FILE
-docker-compose up -d postgres
+if [[ -z $dock ]]; then
+  docker-compose down -v
+  unset PG_CONF_FILE
+  docker-compose up -d postgres
+fi
